requirements.txt
streamlit
langserve
langchain
python-dotenv
fastapi
uvicorn
langchain_community
sse_starlette
beautifulsoup4
pypdf
chromadb
faiss-cpu